fix(BlockBasis): migrate state representation int -> long (unblocks N >= 31)
BlockBasis.PopcountStates used for (int x = 0; x < (1 << N); x++) — the
1 << N int32 expression overflows at N = 31 (to int.MinValue, empty basis)
and wraps at N = 32, 33. A latent bug, never hit until the inverse-iteration
speed-up made the c=2 block reachable at N >= 31.

States are now long throughout: BlockBasis (StatesP/StatesQ, the index
dictionaries, IndexP/IndexQ/FlatIndex parameters, the 1L << N loop bound)
and every consumer that touches a state value — BlockLDecomposition (all
three build paths + BondFlipTargets), SpatialSumKernel, HdSubspaceProjector,
C2ChannelUniformAnalytical, HdChannelBasis, BondHdChannelWeights,
JwBlockBasis (incl. SingleExcitationSite/TwoExcitationSites), and
C2FullBlockSigmaAnatomy.BuildSiteOverlapIndices. Index values (FlatIndex
results, Mp/Mq/MTotal) stay int. Popcount casts widened (uint) -> (ulong),
masks 1 << -> 1L <<.

The 2^N brute-force scan in PopcountStates is unchanged — at N = 33 it adds
tens of seconds, dwarfed by the LU work; combinatorial enumeration is a
possible future perf improvement, not a correctness need.

Full Core.Tests suite green (3674 passed) — the migration round-trips
bit-exact at small N.

+++ compute/RCPsiSquared.Core.Tests/CoherenceBlocks/BlockBasisTests.cs
@@ -23,9 +23,9 @@ public class BlockBasisTests
     {
         var basis = new BlockBasis(N: 5, n: 2);
         Assert.All(basis.StatesP, s =>
-            Assert.Equal(2, System.Numerics.BitOperations.PopCount((uint)s)));
+            Assert.Equal(2, System.Numerics.BitOperations.PopCount((ulong)s)));
         Assert.All(basis.StatesQ, s =>
-            Assert.Equal(3, System.Numerics.BitOperations.PopCount((uint)s)));
+            Assert.Equal(3, System.Numerics.BitOperations.PopCount((ulong)s)));
     }
 
     [Fact]
@@ -44,8 +44,8 @@ public class BlockBasisTests
         var basis = new BlockBasis(N: 5, n: 1);
         // For (p=1, q=3) at N=5: p has popcount 1 (bit 0 set, big-endian → site 4),
         // q has popcount 2. FlatIndex = IndexP(1)·Mq + IndexQ(3).
-        int p = basis.StatesP[2]; // some valid state
-        int q = basis.StatesQ[5];
+        long p = basis.StatesP[2]; // some valid state
+        long q = basis.StatesQ[5];
         int idx = basis.FlatIndex(p, q);
         Assert.Equal(basis.IndexP(p) * basis.Mq + basis.IndexQ(q), idx);
         Assert.InRange(idx, 0, basis.MTotal - 1);
